+++ kubemarine/patches/p1_managed_sandbox_image.py
